fix wasm test
