+++ packages/fuels-accounts/src/provider.rs
@@ -25,8 +25,7 @@ use fuels_core::{
         message::Message,
         message_proof::MessageProof,
         node_info::NodeInfo,
-        transaction::Transaction,
-        transaction_builders::NetworkInfo,
+        transaction::{NetworkInfo, Transaction},
         transaction_response::TransactionResponse,
         tx_status::TxStatus,
     },
